requirements.txt
boto3
click
PyYAML
pytest
moto